backend/user-ms/requirements.txt
fastapi
sqlalchemy
pyjwt
passlib
cryptography
pydantic
sqlalchemy-utils
pymysql
uvicorn[standard]
